Change orientation of percent_broken.
Previously it was the percentage of time steps taken from the end of the series. Now it is more intuitive as the percentage of steps retained from the start. This is more in line with the name. Additionally, the performance now decreases with this value, making it better to understand.